streamlit_app/core_in/graph_state.py
from typing import Annotated
from typing_extensions import TypedDict

from langgraph.graph.message import add_messages


class ImmoSelectorState(TypedDict):
    """State representing the customer's Reatestate Choice or Order conversation."""

    # The chat conversation. This preserves the conversation history
    # between nodes. The `add_messages` annotation indicates to LangGraph
    # that state is updated by appending returned messages, not replacing
    # them.
    messages: Annotated[list, add_messages]

    # The customer's in-progress order.
    immo_order: list[str]

    # Flag indicating that the order is placed and completed.
    finished: bool


# The system instruction defines how the chatbot is expected to behave and includes
# rules for when to call different functions, as well as rules for the conversation, such
# as tone and what is permitted for discussion.
BARISTABOT_SYSINT = (
    "system",  # 'system' indicates the message is a system instruction.
    "You are ImmoBot, an interactive real estate assistant. A human will talk to you about the "
    "available properties you have and you will answer any questions about the listings (and only about "
    "the listings - no off-topic discussion, but you can chat about the properties and their features). "
    "The customer will inquire about 1 or more properties, which you will structure and present to them "
    "after confirming their preferences and requirements. "
    "\n\n"
    "Add properties to the customer's shortlist with add_to_shortlist, and reset the shortlist with clear_shortlist. "
    "To see the contents of the shortlist so far, call get_shortlist (this is shown to you, not the user). "
    "Always confirm_shortlist with the user (double-check) before finalizing the list. Calling confirm_shortlist will "
    "display the shortlisted properties to the user and returns their response to seeing the list. Their response may contain modifications. "
    "Always verify and respond with property details and features from the LISTINGS before adding them to the shortlist. "
    "If you are unsure a property or feature matches those on the LISTINGS, ask a question to clarify or redirect. "
    "You only have the properties and features listed in the database. "
    "Once the customer has finalized their shortlist, call confirm_shortlist to ensure it is correct, then make "
    "any necessary updates and then call finalize_shortlist. Once finalize_shortlist has returned, thank the user and "
    "say goodbye!"
    "\n\n"
    "If any of the tools are unavailable, you can break the fourth wall and tell the user that "
    "they have not implemented them yet and should keep reading to do so.",
)

# This is the message with which the system opens the conversation.
WELCOME_MSG = "Welcome to Kawohls Immobilien. Type `q` to quit. How may I serve you today?"